Fix license formatting so it will be recognized (#78)
GitHub uses the licensee ruby gem to detect the license type of the repository. licensee uses the license texts from choosealicense.com so if the text of your license file differs from their license it will not be recognized. GitHub allows filtering searches by license type and also shows the license type on the homepage of your repository and when viewing the license page but these features are only available when the license is recognized.