cmake and make using same env with CC and CXX (#258)
Each "run:" call gets a new environment.

With this, the first "run:" is exporting CC and CXX.
In this "run:" the cmake stuff is created.

The next "run:" to make, is using the default CC and CXX, which is
GCC instead of clang.

And then some of the stuff is build with clang other with gcc

* moved cmake, make and running tests in one "run:" command, so they
are sharing the same environment

* fixed indention

* renamed the name to Build and run

* Build build-clang-9-Linux and build-clang-10-osx are similar now.

Co-authored-by: finger42 <finger42>