Use a different workaround for problems with IARCC.
This approach may be better as is uses the same syntax using `< ... >` as before. Has been suggested here: https://github.com/ETLCPP/etl/pull/593#issuecomment-1239222424

This is based on the insight, that the problem with IARCC occurs only in case the `__has_include` is used as part of a preprocessor definition. See https://github.com/ETLCPP/etl/pull/593#issuecomment-1239208693