Fix doc comment for exception::line_number()
The documentation comment for `exception::line_number()` in
`include/etl/exception.h` states falsely that the return type for
the function is a `const char*`.

It should state that the return type is `numeric_type`,
which is an `int`.

So change the documentation comment to reflect that the return type is
an `int`.